Extract HTML cleaning into separate class
